refactor: split api tests into three modules
